added reduceLeft scorer; added docs
this is just an exercise in how to do this scoring in multiple ways. The reduce left is demonstrated in this commit. The tailrec version may be more flexible but depending on performance requirements, may be less desirable than an optimized reduce… though at this point, the function is performing pretty well on large lists

Signed-off-by: Larry Mills-Gahl <elemgee@gmail.com>